Move plotting functions into separate module
