+++ src/prfmodel/stimuli/__init__.py
@@ -8,8 +8,6 @@ field (pRF) models and connective field (CF) models are implemented.
 from ._cf import CFStimulus
 from ._prf import GridDimensionsError
 from ._prf import PRFStimulus
-from ._prf import animate_2d_prf_stimulus
-from ._prf import plot_2d_prf_stimulus
 from .base import Stimulus
 
 __all__ = [
@@ -17,6 +15,4 @@ __all__ = [
     "GridDimensionsError",
     "PRFStimulus",
     "Stimulus",
-    "animate_2d_prf_stimulus",
-    "plot_2d_prf_stimulus",
 ]
